Tictactoe.py initial version
Commited todays work. This is the initial version of the tictactoe-game. There are several issues that has to be handled before this is ready, but it is a working prototype and you can play the game, win or lose and replay.
Currently its only player-vs-player and several functionalities from our requirements are yet to be implemented, more details regarding what work needs to be done can be found in the comments of the code.

If you're testing this know that it is so far adapted for python 2.x, some slight changes might be required for it to run with python 3. Also the user input isn't sanities or checked for validity in some cases, so you can easily crash the program if you don't input things exactly right. This is also mentioned in the TODOs.